requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=2.2.0
pillow
numpy
matplotlib
seaborn
scikit-learn
tqdm
